requirements.txt
paho-mqtt==1.5.1
schedule==1.1.0
